Fixed the bug that was causing the tests for bug #2991 to fail.  It was the 'singleton' method I was using to create the class for the tests.  It was caching a bad one.

+++ test/plugins/database/MysqlDriverTest.php
@@ -67,24 +67,32 @@ class MysqlDriverTest extends PHPUnit_Extensions_Database_TestCase
     */
     protected function setUp()
     {
-        $this->pdo = PHPUnit_Util_PDO::factory("mysql://test:test@localhost/test");
-        if (!is_object($this->pdo)) {
+        $this->skipPDOTests = false;
+        try {
+            $this->pdo = PHPUnit_Util_PDO::factory("mysql://test:test@localhost/test");
+            $this->pdo->query("DROP TABLE IF EXISTS `myTable`");
+            $this->pdo->query(
+                "CREATE TABLE `myTable` ("
+                ." `id` int(11) PRIMARY KEY NOT NULL,"
+                ." `name` varchar(32) NOT NULL,"
+                ." `value` float NULL"
+                ." ) TABLESPACE MEMORY;"
+            );
+        } catch (PDOException $e) {
             $this->skipPDOTests = true;
             $this->pdo = PHPUnit_Util_PDO::factory("sqlite::memory:");
-        } else {
-            $this->skipPDOTests = false;
+            $this->pdo->query("DROP TABLE IF EXISTS `myTable`");
+            $this->pdo->query(
+                "CREATE TABLE `myTable` ("
+                ." `id` int(11) PRIMARY KEY NOT NULL,"
+                ." `name` varchar(32) NOT NULL,"
+                ." `value` float NULL"
+                ." ) "
+            );
         }
-        $this->pdo->query("DROP TABLE IF EXISTS `myTable`");
-        $this->pdo->query(
-            "CREATE TABLE IF NOT EXISTS `myTable` ("
-            ." `id` int(11) PRIMARY KEY NOT NULL,"
-            ." `name` varchar(32) NOT NULL,"
-            ." `value` float NULL"
-            ." ) TABLESPACE MEMORY;"
-        );
         parent::setUp();
         $this->table = new DummyTableContainer();
-        $this->o = MysqlDriver::singleton($this->table, $this->pdo);
+        $this->o = new MysqlDriver($this->table, $this->pdo);
     }
 
     /**
@@ -97,7 +105,9 @@ class MysqlDriverTest extends PHPUnit_Extensions_Database_TestCase
     */
     protected function tearDown()
     {
-        $this->pdo->query("DROP TABLE IF EXISTS `myTable`");
+	if (is_object($this->pdo)) {
+            $this->pdo->query("DROP TABLE IF EXISTS `myTable`");
+        }
         unset($this->o);
         unset($this->pdo);
     }
@@ -181,7 +191,7 @@ class MysqlDriverTest extends PHPUnit_Extensions_Database_TestCase
     {
         $this->pdo->query($preload);
         $cols = $this->o->columns();
-        if ($this->SkipPDOTests) {
+        if ($this->skipPDOTests) {
             $this->markTestSkipped("No MySQL server available");
         } else {
             $this->assertSame($expect, $cols);
